requirements.txt
PyPDF2==3.0.1
reportlab==4.0.8
openai>=1.3.5
python-dotenv
pdfplumber>=0.10.0
requests>=2.31.0
streamlit>=1.28.0
python-docx>=0.8.11